Implement double elimination bracket
Focus on implementing the double elimination bracket UI using the provided data structure. Override or remove the single elimination logic as requested.

+++ src/components/TournamentBracket.tsx
@@ -10,7 +10,7 @@ import './TournamentBracket.css';
 
 const TournamentBracket = () => {
   const [zoomLevel, setZoomLevel] = useState(0.8);
-  const [bracketType, setBracketType] = useState<BracketType>('single');
+  const [bracketType, setBracketType] = useState<BracketType>('double');
   
   const handleZoomIn = () => {
     setZoomLevel(prev => Math.min(prev + 0.2, 2));
